Refactor the code to be more OOP
